+++ packages/ens-sync-script/prepare.sh
@@ -1,5 +1,5 @@
 mkdir -p network-contracts/artifacts/contracts/StreamRegistry/StreamRegistryV5.sol
-mkdir -p network-contracts/artifacts/contracts/chainlinkClient/ENSCacheV2Streamr.sol
+mkdir -p network-contracts/artifacts/contracts/ENS/ENSCacheV2Streamr.sol
 cp ../network-contracts/artifacts/contracts/StreamRegistry/StreamRegistryV5.sol/StreamRegistryV5.json ./network-contracts/artifacts/contracts/StreamRegistry/StreamRegistryV5.sol
-cp -r ../network-contracts/artifacts/contracts/chainlinkClient/ENSCacheV2Streamr.sol/ENSCacheV2Streamr.json ./network-contracts/artifacts/contracts/chainlinkClient/ENSCacheV2Streamr.sol
+cp -r ../network-contracts/artifacts/contracts/ENS/ENSCacheV2Streamr.sol/ENSCacheV2Streamr.json ./network-contracts/artifacts/contracts/ENS/ENSCacheV2Streamr.sol
