+++ packages/network-contracts/.openzeppelin/unknown-3338.json
@@ -363,7 +363,7 @@
             "slot": "201",
             "type": "t_mapping(t_string_memory_ptr,t_address)",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:17"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:17"
           },
           {
             "label": "streamrScript",
@@ -371,7 +371,7 @@
             "slot": "202",
             "type": "t_address",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:18"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:18"
           },
           {
             "label": "streamRegistry",
@@ -379,7 +379,7 @@
             "slot": "203",
             "type": "t_contract(IStreamRegistry)21862",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:19"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:19"
           },
           {
             "label": "ensCacheV1",
@@ -387,7 +387,7 @@
             "slot": "204",
             "type": "t_contract(ENSCacheV1)22220",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:20"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:20"
           }
         ],
         "types": {
