+++ packages/network-contracts/.openzeppelin/unknown-4690.json
@@ -348,7 +348,7 @@
             "slot": "201",
             "type": "t_mapping(t_string_memory_ptr,t_address)",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:17"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:17"
           },
           {
             "label": "streamrScript",
@@ -356,7 +356,7 @@
             "slot": "202",
             "type": "t_address",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:18"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:18"
           },
           {
             "label": "streamRegistry",
@@ -364,7 +364,7 @@
             "slot": "203",
             "type": "t_contract(IStreamRegistry)24359",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:19"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:19"
           },
           {
             "label": "ensCacheV1",
@@ -372,7 +372,7 @@
             "slot": "204",
             "type": "t_contract(ENSCacheV1)24717",
             "contract": "ENSCacheV2Streamr",
-            "src": "contracts/chainlinkClient/ENSCacheV2Streamr.sol:20"
+            "src": "contracts/ENS/ENSCacheV2Streamr.sol:20"
           }
         ],
         "types": {
